requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
SpeechRecognition==3.10.0
pyttsx3==2.90
sounddevice==0.4.6
numpy>=1.26.0
python-dateutil==2.8.2
websockets==11.0.3
pydantic==2.4.2
python-jose==3.3.0
APScheduler==3.10.4
scikit-learn==1.4.0
pandas==2.2.0
python-socketio==5.11.1
aiohttp==3.9.3
jinja2==3.1.3
python-dotenv==1.0.1 